Deleted AddFamilyItem and Renames AddFamily
