+++ src/components/Profile.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import {useState} from 'react';
 import participantData from './data/participantData.js';
-import AddFamilyItem from './AddFamily/AddFamilyItem';
+import AddFamilyItem from './AddFamilyPopup';
 import YouAndYourFamilyCard from './YouAndYourFamilyCard';
 import ContactPersonCard from './ContactPersonCard';
 import CarsAndSeatsCard from './CarsAndSeatsCard.js';
